LayoutTests/fast/scrolling/mac/keyboard-rubberband.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true AsyncOverflowScrollingEnabled=true AsyncFrameScrollingEnabled=true ScrollAnimatorEnabled=true ] -->
<html>
<head>
    <script src="../../../resources/ui-helper.js"></script>
    <script src="../../../resources/js-test.js"></script>
    <meta name="viewport" content="initial-scale=1.5, user-scalable=no">
    <script>
        jsTestIsAsync = true;
        
        async function runTest()
        {
            if (window.eventSender) {
                const delay = ms => new Promise(res => setTimeout(res, ms));
                await UIHelper.startMonitoringWheelEvents();
                await UIHelper.rawKeyDown("upArrow");
                await UIHelper.waitForCondition(() => { return window.pageYOffset < -16 });
                shouldBeTrue('window.pageYOffset < 0');
                finishJSTest();
            }
        }
    </script>
</head>
<body onload="runTest()">
    <div style="height: 5000px;">
    </div>
</body>

</html>
