LayoutTests/fast/scrolling/mac/keyboard-rubberband-expected.txt

PASS window.pageYOffset < 0 is true
PASS successfullyParsed is true

TEST COMPLETE
